pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "binex"
version = "0.1.0"
description = "Debuggable runtime for A2A agents"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk",
    "litellm",
    "fastapi",
    "uvicorn",
    "httpx",
    "pydantic>=2.0",
    "pyyaml",
    "click",
    "aiosqlite",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/binex"]

[project.scripts]
binex = "binex.cli.main:cli"

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
